requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
peewee==3.17.8
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
urllib3==2.3.0
celery==5.2.3
redis
rabbitmq
peewee
python-dotenv
